pyproject.toml
[tool.poetry]
name = "cpf-serverless-python"
version = "0.1.0"
description = ""
authors = ["Julio Okuda <julio.okuda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pycpfcnpj = "^1.8"
pydantic = "^2.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
azure-functions = "^1.23.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^0.23.8"
black = "^25.9.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-fail-under=90"
